Prac1/requirements.txt
grpcio
requests
beautifulsoup4
redis